API requests work, refactoring.

+++ src/main/java/com/filunderscore/destiny4j/api/entities/user/membership/IUserMembership.java
@@ -8,5 +8,5 @@ public interface IUserMembership
 	String getDisplayName();
 	
 	String getBungieGlobalDisplayName();
-	Byte getBungieGlobalDisplayNameCode();
+	Short getBungieGlobalDisplayNameCode();
 }
